requirements.txt
Django==4.1.1
django-rest-knox
django-rest-framework